LISA: run uibench tests using "am instrument"
UiBench workloads that involve touch events such as scrolling need to
be run using the InstrumentationTestRunner, and require the
com.android.uibench.janktests package to work properly. Switch the
UiBench workload to send an "am instrument" command to start tests,
and extend the Workload class's _check_availables() function to check
that all necessary packages are available on the device.

"am instrument" takes the number of iterations as an argument, not a
duration, so pass a very large number of iterations and then cut off
the benchmark after the appropriate duration.

Test: Ran all supported tests using UiBench.run(); confirmed all tests
run and that reasonable gfxinfo data is collected.
Bug: 62224943

Change-Id: I9cc28387415c8b05831028425ac564bc53dbca2f
Signed-off-by: Connor O'Brien <connoro@google.com>