tests/benchmark: Consolidate the benchmark confs
The benchmark conf at the moment is the same for all benchmarks. There's no
point in duplicating them. Provide a default conf and if the individual
benchmark needs to provide something else to override it, then can either be
provided by overriding bm_conf or by setting properties in the object. This
brings down total LOC too.

Change-Id: I643f9d3e37cc93a06e5cdb09cad54f117fac84c9
Signed-off-by: Joel Fernandes <joelaf@google.com>