utils/trace: add cgroup event sanitization
This patch aims to build a cgroup_attach_task data frame that is a combination
of fake devlib events (if available), and real cgroup_task_attach events. We
pass along cgroup information to the trace object to help with this (controller
id to name mapping and cgroup names). Some out of tree kernel patches already
have 'cgroup' field, so we don't add cgroup information to those and just add
controller information.

Change-Id: Ib82aa6492d13d1faf5f64fc45d7dd85c2ca2489a
Signed-off-by: Joel Fernandes <joelaf@google.com>