scripts: add a script for generating notebooks
I put together a simple utility script some time back to generate an
ipython notebook from a python file. Then one could run the following
to generate an HTML file:
ipython nbconvert --to html temp_notebook.ipynb

The idea I had was to run analysis scripts that need graphical output
as a 'notebook' and generate an HTML, but optionally not run it and
generate output on the command line (standard output) if the analysis
script chooses to (for example some (most) analysis scripts are simple
enough for command line output).

Change-Id: Ia057be958157d156409ffdc3d93e6364bd98b1ca
Signed-off-by: Joel Fernandes <joelaf@google.com>