libs/utils: fix the plotting of CPU residency.
Upstream changed 'comm' to 'TaskName', which broke parsing.

CPU residency analysis was not in the integration testing loop, so this
includes some changes so that running experiments/run_uibench_cgroup.py
is sufficient smoke testing for this analysis feature.

If running experiments/run_uibench_cgroup.py, the charts no longer pop
up onscreen. This helps with testing, as well as is more consistent with
other analysis scripts work. (e.g. FrequencyAnalysis). pylab grabs the
output image in the notebook, so the behavior in the .ipynb files is the
same.

Fixes: 68655983

Test: Run experiments/run_uibench_cgroup.py, and verify the charts
produced in the results directory.
Test: Run notebooks/residency/task_residencies_uibench.ipynb and
make sure that the charts appear correctly.

Change-Id: I5c8cc54e9d8fde352c779a9a47c9762bd0dd9905